Add functions to apply a rule to a graph.
The matchRule and/or ruleInMatchSpace function have some bugs that may
or may not be a problem. They require that both the lefthand- and
righthand-side of a rule are strongly connected.
Further more, ruleInMatchSpace will fail if matchRule picks a root node
from the lefthand-side that is not in the righthand-side.